+++ lizard_rainapp/test_calculations.py
@@ -1,34 +1,31 @@
 from datetime import timedelta
 from datetime import datetime
-from datetime import tzinfo
 
 from django.test import TestCase
 from lizard_rainapp.calculations import meter_square_to_km_square
 from lizard_rainapp.calculations import herhalingstijd
 from lizard_rainapp.calculations import moving_sum
 
+import pytz
 import logging
 logger = logging.getLogger(__name__)
 
-ZERO = timedelta(0)
-TZ_OFFSET = 60
-TZ_NAME = 'Test'
+UTC = pytz.timezone('UTC')
 
 
 def generate_values(dt_start, dt_stop, td_step, unit, value):
     """Generate list of values similar to the one returned by fews."""
-    timezone_info = FixedOffset(TZ_OFFSET, TZ_NAME)
 
     # Make tzaware
-    dt_start_tz = dt_start.replace(tzinfo=timezone_info)
-    dt_stop_tz = dt_stop.replace(tzinfo=timezone_info)
+    dt_start_utc = UTC.localize(dt_start)
+    dt_stop_utc = UTC.localize(dt_stop)
 
     # Initial conditions
-    dt = dt_start_tz
+    dt = dt_start_utc
     values = []
 
     # Loop
-    while (dt <= dt_stop_tz):
+    while (dt <= dt_stop_utc):
 
         values.append({
             'unit': unit,
@@ -38,28 +35,6 @@ def generate_values(dt_start, dt_stop, td_step, unit, value):
         dt += td_step
 
     return values
-
-
-class FixedOffset(tzinfo):
-    """Fixed offset in minutes east from UTC.
-
-    From http://docs.python.org/library/datetime.html
-
-    A class building tzinfo objects for fixed-offset time zones.
-    Note that FixedOffset(0, "UTC") is a different way to build a
-    UTC tzinfo object."""
-    def __init__(self, offset, name):
-        self.__offset = timedelta(minutes=offset)
-        self.__name = name
-
-    def utcoffset(self, dt):
-        return self.__offset
-
-    def tzname(self, dt):
-        return self.__name
-
-    def dst(self, dt):
-        return ZERO
 
 
 class ComputeTestSuite(TestCase):
@@ -90,16 +65,15 @@ class ComputeTestSuite(TestCase):
 
         # Make start_date and end_date tz aware, just as in rain_stats from
         # adapter
-        timezone_info = FixedOffset(TZ_OFFSET, TZ_NAME)
-        start_date = start_date.replace(tzinfo=timezone_info)
-        end_date = end_date.replace(tzinfo=timezone_info)
+        start_date = UTC.localize(start_date)
+        end_date = UTC.localize(end_date)
 
         moving_sum_kwargs = {
             'values': generate_values(**generate_values_kwargs),
             'td_window': timedelta(hours=1),
             'td_value': td_step,
-            'start_date': start_date,
-            'end_date': end_date}
+            'start_date_utc': start_date,
+            'end_date_utc': end_date}
 
         # 24 hour data and a 1 hour window, there should be no max_values
         self.assertEqual(len(moving_sum(**moving_sum_kwargs)), 0)